picky-krb: crypto: improve des encryption; format code; fix clippy warnings
