+++ picky-krb/src/crypto/des/mod.rs
@@ -1,5 +1,5 @@
 pub mod des3_cbc_sha1_kd;
-pub mod encryopt;
+pub mod encrypt;
 
 pub const DES3_BLOCK_SIZE: usize = 8;
 pub const DES3_KEY_SIZE: usize = 24;
